# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API для системы поиска и управления историческими хрониками",
        "title": "ChronicleSearch API",
        "contact": {
            "name": "API Support",
            "email": "support@chroniclesearch.ru"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/chronicle_research/{id}/chronicles/{chronicle_id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновление количества, порядка или других полей связи хроники с заявкой (требуется авторизация)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chronicle_research"
                ],
                "summary": "Обновить хронику в заявке",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID хроники",
                        "name": "chronicle_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Обновляемые поля (quantity, order, etc.)",
                        "name": "updates",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаление хроники из заявки (требуется авторизация)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chronicle_research"
                ],
                "summary": "Удалить хронику из заявки",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID хроники",
                        "name": "chronicle_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите токен в формате: Bearer {токен}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}